+++ packages/styled-components/components/text/themes/h3.js
@@ -1,0 +1,9 @@
+/* eslint max-len: 0 */
+import styled from 'styled-components';
+
+// eslint-disable import/prefer-default-export.
+export const TextWrapper = styled.div`
+  font-size: 1.55rem;
+  margin-bottom: .8rem;
+`;
+// eslint-enable.