+++ packages/styled-components/components/text/themes/html.js
@@ -17,7 +17,7 @@ const htmlVars = {
 };
 
 // eslint-disable import/prefer-default-export.
-export const HTMLWrapper = styled.div`
+export const TextWrapper = styled.div`
   address,
   area,
   article,