+++ packages/styled-components/index.js
@@ -2,22 +2,26 @@ import App from 'components/app';
 import Byline from 'components/byline';
 import Container from 'components/container';
 import Fragment from 'components/fragment';
-import HTML from 'components/html';
 import Link from 'components/link';
 import Logo from 'components/logo';
 import Menu from 'components/menu';
+import Pagination from 'components/pagination';
+import SearchForm from 'components/searchForm';
 import SocialSharing from 'components/socialSharing';
+import Text from 'components/text';
 
 export {
   App,
   Byline,
   Container,
   Fragment,
-  HTML,
   Link,
   Logo,
   Menu,
+  Pagination,
+  SearchForm,
   SocialSharing,
+  Text,
 };
 
 const ComponentMap = {
@@ -28,12 +32,13 @@ const ComponentMap = {
   'irving/footer-wrapper': Fragment,
   'irving/fragment': Fragment,
   'irving/header-wrapper': Fragment,
-  'irving/html': HTML,
   'irving/link': Link,
   'irving/logo': Logo,
   'irving/menu': Menu,
+  'irving/pagination': Pagination,
+  'irving/search-form': SearchForm,
   'irving/social-sharing': SocialSharing,
-  'irving/text': Fragment,
+  'irving/text': Text,
   app: App,
 };