+++ CHANGELOG.md
@@ -3,6 +3,48 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.15.0](https://github.com/alleyinteractive/irving/compare/v2.14.0...v2.15.0) (2020-06-16)
+
+**Note:** Version bump only for package irving
+
+
+
+
+
+# [2.14.0](https://github.com/alleyinteractive/irving/compare/v2.14.0-beta.1...v2.14.0) (2020-06-15)
+
+**Note:** Version bump only for package irving
+
+
+
+
+
+# [2.14.0-beta.1](https://github.com/alleyinteractive/irving/compare/v2.14.0-beta.0...v2.14.0-beta.1) (2020-06-15)
+
+**Note:** Version bump only for package irving
+
+
+
+
+
+# [2.14.0-beta.0](https://github.com/alleyinteractive/irving/compare/v2.13.0...v2.14.0-beta.0) (2020-06-15)
+
+
+### Bug Fixes
+
+* **social sharing:** fix path in socialsharingitem.js ([36057e6](https://github.com/alleyinteractive/irving/commit/36057e6a5d74c1a74b1134ee9a3898a309cc6b64))
+
+
+### Features
+
+* **add social sharing component to component map:** add social sharing component to component map ([4bbb27f](https://github.com/alleyinteractive/irving/commit/4bbb27fea40140a255b638ac381ae77bd6a75552))
+* **social sharing:** add social sharing styled component ([f5daaa4](https://github.com/alleyinteractive/irving/commit/f5daaa4a9dc373e3677a381b0052752f980e19a0))
+* **social sharing:** add wp i18n package to dependencies ([94e06f5](https://github.com/alleyinteractive/irving/commit/94e06f5aa9c020e0dddf03e5858aee86f0bb072e))
+
+
+
+
+
 # [2.13.0](https://github.com/alleyinteractive/irving/compare/v2.13.0-rc.0...v2.13.0) (2020-06-12)